downgrade rusqlite dependency from 0.24.2 to 0.23.1 for Buster
The Rust compiler version on Debian 10 ("Buster") does not support
the matches_macro feature used in rusqlite 0.24.2. It fails with

    error[E0658]: use of unstable library feature 'matches_macro'
       --> /root/.cargo/registry/src/github.com-1ecc6299db9ec823/rusqlite-0.24.2/src/types/from_sql.rs:146:44
        |
    146 |         i64::column_result(value).map(|i| !matches!(i, 0))
        |                                            ^^^^^^^
        |
        = note: for more information, see https://github.com/rust-lang/rust/issues/65721
    error: aborting due to previous error
    For more information about this error, try `rustc --explain E0658`.
    error: could not compile `rusqlite`.

when trying to compile rusqlite 0.24.2. However, it does compile
with the older version 0.23.1 of rusqlite. Until there is a more
recent Rust compiler version available in Debian 10, e.g. via the
backports repository, it will probably stay at this version.