Revert "dependencies: update tracing to 0.1.26 + tracing-core to 0.1.18"
This reverts commit 4063a86ed3ebd6ffc66d445868d8f8a5c9504358.
The authors of tracing + tracing-core changed the Minimum
Supported Rust Version (MSRV) from 1.40 to 1.42 with a _patch_
version update. Who does that? That is a breaking change and does
require a _major_ version update in SemVer. Yes, they are still
in 0.x.y, so they do not need to care about version guarantees,
but still that is somehow unexpected and should get a prominent
notification / warning for the users of these crates.

Anyway, back to 0.1.25 and 0.1.17 it is then.