do not validate certificate of health-infobase.canada.ca
This is the official source for Canadian COVID numbers, but the
domain currently has a problem with an incomplete certificate
chain, so any HTTPS requests will fail. To avoid that failure,
the certificate validation for requests to that domain are
currently disabled. Let's hope they fix it and we can re-enable
the validation later.