ci: allow GHA workflow on Windows to fail
Currently, ntapi 0.3.7 has code that generated just a warning in
previous Rust versions. However, since Rust 1.69.0 this code is
a hard error, so the build fails. To allow that failure without
failing the pipeline, a new flag is set.

This change should be reverted as soon as `ntapi` is removed
from the dependency tree.