ci: pin rustc in AppVeyor builds to 1.68.2
Currently, ntapi 0.3.7 has code that generated just a warning in
previous Rust versions. However, since Rust 1.69.0 this code is
a hard error, so the build fails. To allow the build to pass the
rustc version is now pinned to 1.68.2

This change should be reverted as soon as `ntapi` is removed
from the dependency tree.

[ci skip]