rewrite ECDC version of the `db` command for speed + total numbers
The actual speedup may vary, depending on your environment. On my
old notebook the duration of the db command changed from 41 seconds
to 32 seconds for a file with less than one year of data, and from
2 minutes 21 seconds down to 39 seconds for approx. 3.5 years of
data (which is the current size of the CSV file), which is a huge
performance win. :)