ci: remove AppVeyor CI builds and build on GHA instead
AppVeyor builds take longer, because they are running in sequence
and not in parallel as GHA builds do. Furthermore, a separate CI
provider just for a single build target (Windows MSVC toolchain)
seems to be a bit overkill, so the MSVC build now happens on GHA,
although with only one job (just x86_64) instead of two jobs (for
i686 and x86_64).