add data up to 2023-07-26
Something in the WHO's CSV data seems to be messed up, because
some entries occur twice while others vanished. This may not be
the data the WHO intended to publish, but since most of it seems
to be correct we take it. Chances are those errors get fixed next
week when new data is published.

[ci skip]