ci: remove thread limitation of tests on GitLab
The limitation was introduced in commit f9968da57e1a when there
still was a collect operation using the disease.sh API. That API
had some kind of rate limit which caused tests to fail when too
many threads were used. However, the API is not used anymore, so
the limit can be safely removed.